+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/Details.cshtml.cs
@@ -1,34 +1,10 @@
 using DfE.FindInformationAcademiesTrusts.Data;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts;
 
-public class DetailsModel : PageModel, ITrustsAreaModel
+public class DetailsModel : TrustsAreaModel
 {
-    private readonly ITrustProvider _trustProvider;
-
-    public DetailsModel(ITrustProvider trustProvider)
+    public DetailsModel(ITrustProvider trustProvider) : base(trustProvider, "Details")
     {
-        _trustProvider = trustProvider;
-    }
-
-    [BindProperty(SupportsGet = true)] public string Uid { get; set; } = "";
-    public Trust Trust { get; set; } = default!;
-    public string PageName => "Details";
-    public string Section => ViewConstants.AboutTheTrustSectionName;
-
-
-    public async Task<IActionResult> OnGetAsync()
-    {
-        var trust = await _trustProvider.GetTrustByUidAsync(Uid);
-
-        if (trust == null)
-        {
-            return new NotFoundResult();
-        }
-
-        Trust = trust;
-        return Page();
     }
 }
