+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/ITrustsAreaModel.cs
@@ -5,6 +5,14 @@ namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts;
 public interface ITrustsAreaModel
 {
     Trust Trust { get; set; }
-    string PageName { get; }
+
+    /// <summary>
+    /// The name of the page as displayed in the page h1
+    /// </summary>
+    string PageName { get; init; }
+
+    /// <summary>
+    /// The name of the section the page sits under in side navigation
+    /// </summary>
     string Section { get; }
 }
