+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/Overview.cshtml.cs
@@ -1,34 +1,10 @@
 ﻿using DfE.FindInformationAcademiesTrusts.Data;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts;
 
-public class OverviewModel : PageModel, ITrustsAreaModel
+public class OverviewModel : TrustsAreaModel
 {
-    private readonly ITrustProvider _trustProvider;
-
-    public OverviewModel(ITrustProvider trustProvider)
+    public OverviewModel(ITrustProvider trustProvider) : base(trustProvider, "Overview")
     {
-        _trustProvider = trustProvider;
-    }
-
-    [BindProperty(SupportsGet = true)] public string Uid { get; set; } = "";
-    public Trust Trust { get; set; } = default!;
-    public string PageName => "Overview";
-    public string Section => ViewConstants.AboutTheTrustSectionName;
-
-
-    public async Task<IActionResult> OnGetAsync()
-    {
-        var trust = await _trustProvider.GetTrustByUidAsync(Uid);
-
-        if (trust == null)
-        {
-            return new NotFoundResult();
-        }
-
-        Trust = trust;
-        return Page();
     }
 }
